+++ src/DelimitedDataParser/Exporter.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Data;
 using System.IO;
+using System.Linq;
 
 namespace DelimitedDataParser
 {
@@ -15,10 +16,15 @@ namespace DelimitedDataParser
         /// </summary>
         public static readonly char TabSeparator = '\t';
 
+        private static readonly string[] UnsafeLeadingCharacters = { "=", "+", "-", "@" };
+
         private ISet<string> _columnNamesAsText;
         private char _fieldSeparator = ',';
         private bool _includeEscapeCharacters = true;
+        private bool _sanitizeStrings = false;
         private bool _outputColumnHeaders = true;
+        private bool _useExtendedPropertyForColumnName = false;
+        private string _extendedPropertyKey;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="Exporter"/> class.
@@ -28,8 +34,8 @@ namespace DelimitedDataParser
         }
 
         /// <summary>
-        /// The character used as the field delimiter in the output. The default value is <c>,</c>,
-        /// i.e. CSV output.
+        /// The character used as the field delimiter in the output. The default value is
+        /// '<c>,</c>', i.e. CSV input.
         /// </summary>
         public virtual char FieldSeparator
         {
@@ -65,6 +71,23 @@ namespace DelimitedDataParser
         }
 
         /// <summary>
+        /// Specifies whether strings should be sanitized, prepending blacklisted characters at the
+        /// start of the string with a single quote "'". The default value is <c>false</c>.
+        /// </summary>
+        public virtual bool SanitizeStrings
+        {
+            get
+            {
+                return _sanitizeStrings;
+            }
+
+            set
+            {
+                _sanitizeStrings = value;
+            }
+        }
+
+        /// <summary>
         /// Specifies whether an initial row containing column names should be written to the
         /// output. The default value is <c>true</c>.
         /// </summary>
@@ -85,8 +108,8 @@ namespace DelimitedDataParser
         /// Clear all "columns as text" settings.
         /// </summary>
         /// <remarks>
-        /// Calling this method clears any "columns as text" settings set via the <see
-        /// cref="SetColumnsAsText(IEnumerable{string})"/> method.
+        /// Calling this method clears any "columns as text" settings set via the
+        /// <see cref="SetColumnsAsText(IEnumerable{string})"/> method.
         /// </remarks>
         public virtual void ClearColumnsAsText()
         {
@@ -94,7 +117,40 @@ namespace DelimitedDataParser
         }
 
         /// <summary>
-        /// Write the input <paramref name="DataTable"/> to the specified <see cref="TextWriter"/>.
+        /// Clear the extended property key setting.
+        /// </summary>
+        /// <remarks>
+        /// Calling this method clears the extended property key setting set via the
+        /// the <see cref="UseExtendedPropertyForColumnName(string)"/> method.
+        /// </remarks>
+        public virtual void ClearExtendedPropertyForColumnName()
+        {
+            _useExtendedPropertyForColumnName = false;
+            _extendedPropertyKey = default(string);
+        }
+
+        /// <summary>
+        /// Populates column headers using the value stored on DataColumn.ExtendedProperties.
+        /// </summary>
+        /// <remarks>
+        /// If no ExtendedProperty can be found that matches the key, the default ColumnName will be used.
+        /// </remarks>
+        /// <param name="key">
+        /// The key that the ExtendedProperties value is stored under.
+        /// </param>
+        public virtual void UseExtendedPropertyForColumnName(string key)
+        {
+            if (string.IsNullOrEmpty(key))
+            {
+                throw new ArgumentNullException("key");
+            }
+
+            _useExtendedPropertyForColumnName = true;
+            _extendedPropertyKey = key;
+        }
+
+        /// <summary>
+        /// Write the input <paramref name="dataTable"/> to the specified <see cref="TextWriter"/>.
         /// </summary>
         /// <param name="dataTable">The <see cref="DataTable"/> containing the data to export.</param>
         /// <param name="writer">The <see cref="TextWriter"/> to be written to.</param>
@@ -115,12 +171,6 @@ namespace DelimitedDataParser
             if (writer == null)
             {
                 throw new ArgumentNullException("writer");
-            }
-
-            // Ensure escape characters are included unless we are exporting tab separated values
-            if (!_includeEscapeCharacters && _fieldSeparator != TabSeparator)
-            {
-                throw new InvalidOperationException();
             }
 
             if (dataTable.Columns.Count > 0)
@@ -165,42 +215,60 @@ namespace DelimitedDataParser
         }
 
         /// <summary>
-        /// Escape the input <paramref name="String"/> for use in a CSV.
+        /// Escape the input <paramref name="value"/> for use in a CSV.
         /// </summary>
-        /// <param name="value">The <see cref="String"/> to be escaped.</param>
+        /// <param name="value">The <see cref="string"/> to be escaped.</param>
         /// <param name="valueAsText">
-        /// Whether the input <paramref name="String"/> should be treated as a text column.
+        /// Whether the input <paramref name="value"/> should be treated as a text column.
         /// </param>
         /// <returns>The escaped string</returns>
         private string CsvEscape(string value, bool valueAsText)
         {
-            if (!_includeEscapeCharacters)
+            if (_includeEscapeCharacters)
             {
-                return value;
+                value = value.Replace(@"""", @"""""");
+
+                if (valueAsText)
+                {
+                    value = string.Concat(@"""=""""", value, @"""""""");
+                }
+                else
+                {
+                    value = string.Concat(@"""", value, @"""");
+                }
             }
 
-            value = value.Replace(@"""", @"""""");
-
-            if (valueAsText)
+            if (_sanitizeStrings)
             {
-                value = string.Concat(@"""=""""", value, @"""""""");
-            }
-            else
-            {
-                value = string.Concat(@"""", value, @"""");
+                value = Sanitize(value);
             }
 
             return value;
         }
 
         /// <summary>
-        /// Whether the input <paramref name="DataColumn"/> should be treated as a text column.
+        /// Sanitize the input <paramref name="value"/> for use in a CSV.
+        /// </summary>
+        /// <param name="value">The <see cref="string"/> to be sanitized.</param>
+        /// <returns>The sanitized string</returns>
+        private string Sanitize(string value)
+        {
+            if (value != null && UnsafeLeadingCharacters.Any(value.StartsWith))
+            {
+                return string.Concat("'", value);
+            }
+
+            return value;
+        }
+
+        /// <summary>
+        /// Whether the input <paramref name="column"/> should be treated as a text column.
         /// </summary>
         /// <param name="column">The <see cref="DataColumn"/> to be checked.</param>
         /// <returns>
-        /// A <see cref="Boolean"/> specifying whether the column should be treated as a text column.
+        /// A <see cref="bool"/> specifying whether the column should be treated as a text column.
         /// </returns>
-        /// <exception cref="ArgumentNullException"><paramref name="textReader"/> is null.</exception>
+        /// <exception cref="ArgumentNullException"><paramref name="column"/> is null.</exception>
         private bool GetIsColumnAsText(DataColumn column)
         {
             if (column == null)
@@ -217,7 +285,7 @@ namespace DelimitedDataParser
         }
 
         /// <summary>
-        /// Write an initial row containing the column names from <paramref name="DataTable"/> to
+        /// Write an initial row containing the column names from <paramref name="dataTable"/> to
         /// the specified <see cref="TextWriter"/>.
         /// </summary>
         /// <param name="dataTable">The <see cref="DataTable"/> containing the columns to be written.</param>
@@ -233,12 +301,19 @@ namespace DelimitedDataParser
                     writer.Write(_fieldSeparator);
                 }
 
-                writer.Write(CsvEscape(col.ColumnName, false));
+                var columnName = col.ColumnName;
+
+                if (_useExtendedPropertyForColumnName && col.ExtendedProperties.ContainsKey(_extendedPropertyKey))
+                {
+                    columnName = col.ExtendedProperties[_extendedPropertyKey].ToString();
+                }
+
+                writer.Write(CsvEscape(columnName, false));
             }
         }
 
         /// <summary>
-        /// Write the <paramref name="DataRow"/> from the <paramref name="DataTable"/> to the
+        /// Write the <paramref name="row"/> from the <paramref name="dataTable"/> to the
         /// specified <see cref="TextWriter"/>.
         /// </summary>
         /// <param name="dataTable">The <see cref="DataTable"/> containing the columns to export.</param>