+++ src/DelimitedDataParser/Properties/AssemblyInfo.cs
@@ -6,4 +6,4 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyProduct("Delimited Data Parser")]
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(true)]
-[assembly: AssemblyVersion("3.0.*")]
+[assembly: AssemblyVersion("3.3.*")]