+++ test/DelimitedDataParser.Test/Properties/AssemblyInfo.cs
@@ -5,4 +5,4 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTitle("Delimited Data Parser Tests")]
 [assembly: AssemblyProduct("Delimited Data Parser")]
 [assembly: ComVisible(false)]
-[assembly: AssemblyVersion("3.0.*")]
+[assembly: AssemblyVersion("3.3.*")]